WIP Feat: Basic User registration page && Login Page. Token is stored on localstorage and into a variable, User info stored onto a variable, both with UserContext. There are still some visual issues, mobile responsive still not ok on both pages, font family, submit button is displaced to the bottom when an error apperars on an Input field, missing links, need to rename some placeholders so everything is in spanish. Validation schema is missing in Login. Functionality is ok
