--- src/components/RegistrationForm/BasicRegistrationForm.css
@@ -1,74 +1,0 @@
-body {
-    background-image: url(../assets/backgroundImages/pexels-luis-del-río-15286.jpg) ;
-}
-.box-container {
-    position: absolute;
-    top: 210px;
-    right: 150px;
-}
-
-form {
-    width: 95%;
-    max-width: 1000px;
-    min-width: 500px;
-    margin: 0 auto;
-  }
-
-  .form-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 15px;
-  }
-
-  input,
-  select {
-    width: 100%;
-    padding: 1rem 1rem;
-    font-size: 1rem;
-    color: white;
-    border: 2px solid #4a5568;
-    background-color: #2d3748;
-    border-radius: 10px;
-    outline: none;
-    margin: 10px;
-  }
-
-  input:focus,
-  select:focus {
-    border-color: #4299e1;
-  }
-  
-  input::placeholder,
-  select::placeholder {
-    color: #a0aec0;
-  }
-
-  button {
-    display: block;
-    margin: 1rem 0;
-    padding: 0.35rem 0.5rem;
-    background-color: #42B7D0;
-    color: #FFFFFF;
-    border: none;
-    border-radius: 5px;
-    width: 170px;
-    height: 35px;
-    font-size: 1rem;
-    font-weight: bold;
-    cursor: pointer;
-    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
-  }
-
-  .button-container {
-    margin-top: 10px;
-  }
-
-  input.input-error,
-  select.input-error {
-    border-color: #fc8181;
-  }
-  
-  button:disabled {
-    opacity: 0.35;
-  }
