+++ src/pages/BasicRegistration/BasicRegistrationPage.css
@@ -1,0 +1,129 @@
+
+.basic {
+    background-image: url(../../components/assets/backgroundImages/pexels-luis-del-río-15286.jpg) ;
+    background-size: cover;
+    background-repeat: no-repeat;
+    height: 100vh;
+    width: 100vw;
+  }
+  
+.logo-container {
+    position: absolute;
+    height: 96px;
+    width: 443px;
+    left: 200px;
+    top: 436px;
+}
+
+svg path {
+    fill: #FFFFFF
+}
+
+svg polygon {
+    fill: #FFFFFF;
+}
+
+svg rect {
+    fill: #FFFFFF;
+}
+
+.box-container-basic-registration {
+    position: absolute;
+    align-items: center;
+    top: 30vh;
+    right: 15vw;
+    left: auto;
+    bottom: auto;
+}
+
+.basic-registration-title {
+    font-style: normal;
+    font-weight: 700;
+    font-size: 36px;
+    line-height: 47px;
+    margin: 0px;
+  
+  }
+  
+  .basic-registration-text {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 27px;
+    margin-top: 5px;
+  }
+
+  .basic-registration-text-link {
+    color: blue;
+  }
+  
+  .box-layout-basic-registration {
+    padding: 40px;
+  }
+
+  .form-container-basic-registration {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 15px;
+    width: 95%;
+    max-width: 1000px;
+    margin: 0 auto;
+  }
+
+  input,
+  select {
+    width: 100%;
+    padding: 1rem 1rem;
+    font-size: 1rem;
+    color: white;
+    border: 2px solid #4a5568;
+    background-color: #2d3748;
+    border-radius: 10px;
+    outline: none;
+    margin: 10px;
+  }
+
+  input:focus,
+  select:focus {
+    border-color: #4299e1;
+  }
+  
+  input::placeholder,
+  select::placeholder {
+    color: #a0aec0;
+  }
+
+  button {
+    display: block;
+    margin: 1rem 0;
+    padding: 0.35rem 0.5rem;
+    background-color: #42B7D0;
+    color: #FFFFFF;
+    border: none;
+    border-radius: 5px;
+    width: 170px;
+    height: 35px;
+    font-size: 1rem;
+    font-weight: bold;
+    cursor: pointer;
+    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
+  }
+
+
+  input.input-error,
+  select.input-error {
+    border-color: #fc8181;
+  }
+  
+  button:disabled {
+    opacity: 0.35;
+  }
+
+  .error {
+    color: #fc8181;
+    font-size: 0.75rem;
+    text-align: left;
+    margin-top: 0.25rem;
+  }
