--- src/pages/BasicRegistrationPage.css
@@ -1,19 +1,0 @@
-.logo-container {
-    position: absolute;
-    height: 96px;
-    width: 443px;
-    left: 200px;
-    top: 436px;
-}
-
-svg path {
-    fill: #FFFFFF
-}
-
-svg polygon {
-    fill: #FFFFFF;
-}
-
-svg rect {
-    fill: #FFFFFF;
-}
